Revert "Revert "Merge branch 'foryasee_rework' into Rework-1.0.0""
This reverts commit ce687da475875f41e9dd68fc133042ef009fd332.

+++ src/main/java/fun/rubicon/commands/general/CommandPremium.java
@@ -64,7 +64,7 @@ public class CommandPremium extends CommandHandler {
                 if ((author.isPremium())) {
                     return message(EmbedUtil.info(invocation.translate("command.premium.already.title"), invocation.translate("command.premium.until")+" " + author.formatExpiryDate().replace("%","")));
                 }
-                long userMoney = author.getMoney();
+                int userMoney = author.getMoney();
                 if (userMoney - PriceList.PREMIUM.getPrice() >= 0) {
                     author.setPremium(PREMIUM_TIME);
                     author.setMoney(userMoney - PriceList.PREMIUM.getPrice());
